Rewrite CheckFiles and commit hook
The common checks are now shared between the scripts, and the common
code in the checks is factored out.

The code is slower (mostly CheckFiles) due to processing each file
separately instead of batch processing with grep, but since CheckFiles
is not frequently run interactively this should not be a big deal.

There are now unit tests for the checks.  They are not run with the
main tests because we don't have a good way to test helper scripts.