Improve latex support in Doxygen
We now are able to have equation numbers and refer to equations using
normal LaTeX commands. Inline math can be triggered using ^^ instead
of \f$ to make writing latex in Doxygen easier. A python script that
will do most of the work for converting a latex file to MathJAX
compatible code is provided. This should be used to convert existing
files to Doxygen, so we don't have to have lots of random tex files
scattered throughout the repository.