Have using std::abs in DataVector.hpp
It is implementation define whether abs in the global namespace is the
same as std::abs (overloads for doubles, etc.), which means we want to
use std::abs (or fabs) on doubles, but cannot use std::abs on
DataVector. This can be remedied by having a 'using std::abs' in the
file.