UpdateU no longer removes dt_vars from box
The temporal derivative is no longer removed from the box so that it
can be mutated later. The reason for this change is that later we will
re-use the oldest history data in the time stepper history to avoid
having to re-allocate data the temporal derivatives. Currently the
time steppers do not support plundering the history in such a fashion
so the allocation is currently still necessary.