Add Rational class
This class is much faster than boost::rational, probably because the
latter is templated on the integer type and therefore can't make many
assumptions or optimizations (and optimized specializations are not
implemented).  Mathematical operators average about twice as fast,
albeit with wide variation.  Ordering operators are roughly eight
times as fast.