Improve defaults for GH constraint damping parameters
This change is based on a set of tests with different
profiles for both Gamma0 and Gamma2. evolving single
black hole spacetimes with Bjorhus boundary conditions.
In addition to the primary Gaussian profile for these
parameters (centered on the hold), these tests suggested
that adding a second Gaussian profile to Gamma0 and Gamma2
that is slower decaying (as a function of radius) and is
2 orders of magnitude smaller in amplitude compared
to the central profile, can make constraints very stable.